requirements.txt
PyPDF2==3.0.1
watchdog==3.0.0
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.0
python-docx==1.0.1
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
darkdetect==0.8.0 